Simplified the NS_* constants, the HELLO timeout computation and fixed some ns_* parameter issues.

+++ src/hash.h
@@ -3,7 +3,7 @@
 
 #include <glib.h>
 
-void ns_hash_table_insert(GHashTable *table, int id, const char *name);
+void ns_hash_table_insert(GHashTable *table, unsigned short id, const char *name);
 void ns_hash_table_free(gpointer key, gpointer value, gpointer user_data);
 gboolean ns_hash_table_check_last_seen(gpointer key, gpointer value, gpointer user_data);
